+++ R/print.text.R
@@ -1,0 +1,2 @@
+
+print.text <- function(x, ...) cat(paste0(x, collapse = "\n"))